net/test: extract mock DNS server into a shared helper
Move the wire-format reply builder and UDP server out of the liveness
test so upcoming resolver tests can reuse them, and teach the server two
more modes: answering A queries with a fixed record, and blackholing
(receive, never reply) to simulate a resolver that loses queries.
